Finish docs README.md. Remove cvat related changes
